version 0.91 - added score calculation
