+++ README.md
@@ -1,2 +1,6 @@
 # kosteczki
+**Version. 0.91**
+- added score calculation
+- modified next_player
+- moved some code to mechanics
 my original board game moved to pygame for testing. Requiers pygame lib
